#61 인덱스 페이지 - 배너 컨테이너 인자로 넘기도록 수정 - 인덱스 페이지 작성
