+++ frontend/components/common/ContentLayout/ContentLayout.scss
@@ -1,5 +1,5 @@
 .content-layout {
-  .content-item:not(:first-child) {
-    margin-top: 12px;
+  .content-item {
+    padding-top: 12px;
   }
 }
