+++ frontend/components/common/ContentLayout/ContentLayout.tsx
@@ -6,13 +6,15 @@ const cx = classNames.bind(styles);
 
 interface IProps {
   children: any;
+  Banner?: any;
 }
 
 class ContentLayout extends Component<IProps> {
   public render() {
-    const { children } = this.props;
+    const { children, Banner } = this.props;
     return (
       <div className={cx('content-layout')}>
+        {Banner && <Banner />}
         {React.Children.map(children, child => (
           <div className={cx('content-item')}>{child}</div>
         ))}
